src/main/java/electronics/Ordinateur.java
package electronics;

import java.util.ArrayList;

/**
 * Décrivez votre classe Ordinateur ici.
 *
 * @author (votre nom)
 * @version (un numéro de version ou une date)
 */
public class Ordinateur
{
	public String type;
	public String brand;
	public ArrayList<Peripherique> peripheriques;

	public String getType(){
        return this.type;
    }
    
    public void setType(String type){
        this.type = type;
    }
    
    public String getBrand(){
        return this.brand;
    }
    
    public void setBrand(String brand){
        this.brand = brand;
    }
    
    public ArrayList<Peripherique> getPeripheriques() {
		return peripheriques;
	}

	public void setPeripheriques(ArrayList<Peripherique> peripheriques) {
		this.peripheriques = peripheriques;
	}

    /**
     * Constructeur d'objets de classe Ordinateur
     */
    public Ordinateur(String type, String brand)
    {
        // initialisation des variables d'instance
        this.type = type;
        this.brand = brand;
        this.peripheriques = new ArrayList<Peripherique>();
    }



	/**
     * Un exemple de méthode - remplacez ce commentaire par le vôtre
     *
     * @param  y   le paramètre de la méthode
     * @return     la somme de x et de y
     */
    public String sampleMethod()
    {
        // Insérez votre code ici
    	String peripheriques = "\nPeripheriques : \n";
    	for(Peripherique p : this.peripheriques) {
    		peripheriques = peripheriques+p.getType()+"\n";
    	}
        return type + " " + brand + peripheriques;
    }
    
    public void addPeripherique(Peripherique p) {
    	p.setOrdi(this);
    	this.peripheriques.add(p);
    }

	@Override
	public boolean equals(Object obj) {
	
		 if (obj instanceof Ordinateur) {
		        Ordinateur ordi= (Ordinateur)obj;
		         return ordi.getType().equals(this.type) && ordi.getBrand().equals(this.brand) 
		        		 && ordi.getPeripheriques().equals(this.peripheriques);
		          
		    }
		    return false;
		
	}
    
    
}
